Add method to dump and modify data
This allows you to get and set the key that self.items is set to, as well as adds a get items getter to iterate over the items in a key. Lastly this updates the tutorial keys to resemble all the possible tutorials and adds a method to disable the entire tutorial on first startup.